+++ libs/access-control/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod access;
 pub mod act;
 mod adapter;
+pub mod collab;
 pub mod enforcer;
 pub mod metrics;
 mod request;
+pub mod workspace;
