+++ services/appflowy-collaborate/src/collab/mod.rs
@@ -1,3 +1,4 @@
+pub mod access_control;
 pub mod cache;
 pub mod disk_cache;
 pub mod mem_cache;
