+++ services/appflowy-collaborate/src/group/cmd.rs
@@ -1,18 +1,20 @@
-use crate::error::RealtimeError;
-use crate::group::manager::GroupManager;
-use crate::RealtimeAccessControl;
+use std::sync::Arc;
+
 use async_stream::stream;
+use collab::entity::EncodedCollab;
+use dashmap::DashMap;
+use futures_util::StreamExt;
+use tracing::{instrument, trace, warn};
+
+use access_control::collab::RealtimeAccessControl;
 use collab_rt_entity::user::RealtimeUser;
 use collab_rt_entity::{ClientCollabMessage, ServerCollabMessage, SinkMessage};
 use collab_rt_entity::{CollabAck, RealtimeMessage};
-use dashmap::DashMap;
 use database::collab::CollabStorage;
 
 use crate::client::client_msg_router::ClientMessageRouter;
-use collab::entity::EncodedCollab;
-use futures_util::StreamExt;
-use std::sync::Arc;
-use tracing::{instrument, trace, warn};
+use crate::error::RealtimeError;
+use crate::group::manager::GroupManager;
 
 /// Using [GroupCommand] to interact with the group
 /// - HandleClientCollabMessage: Handle the client message
