--- services/realtime/src/main.rs
@@ -1,3 +1,0 @@
-fn main() {
-  println!("Hello, world!");
-}
