+++ src/biz/actix_ws/server/rt_actor.rs
@@ -1,12 +1,16 @@
-use crate::biz::actix_ws::client::rt_client::{RealtimeClientWebsocketSinkImpl, RealtimeServer};
-use crate::biz::actix_ws::entities::{ClientMessage, ClientStreamMessage, Connect, Disconnect};
+use std::ops::Deref;
+
 use actix::{Actor, Context, Handler, ResponseFuture};
+use tracing::{error, info, warn};
+
+use access_control::collab::RealtimeAccessControl;
 use appflowy_collaborate::error::RealtimeError;
-use appflowy_collaborate::{CollaborationServer, RealtimeAccessControl};
+use appflowy_collaborate::CollaborationServer;
 use collab_rt_entity::user::UserDevice;
 use database::collab::CollabStorage;
-use std::ops::Deref;
-use tracing::{error, info, warn};
+
+use crate::biz::actix_ws::client::rt_client::{RealtimeClientWebsocketSinkImpl, RealtimeServer};
+use crate::biz::actix_ws::entities::{ClientMessage, ClientStreamMessage, Connect, Disconnect};
 
 #[derive(Clone)]
 pub struct RealtimeServerActor<S, AC>(pub CollaborationServer<S, AC>);
