+++ src/biz/casbin/mod.rs
@@ -1,7 +1,4 @@
-mod collab_ac;
 pub mod pg_listen;
 mod workspace_ac;
 
-pub use collab_ac::CollabAccessControlImpl;
-pub use collab_ac::RealtimeCollabAccessControlImpl;
 pub use workspace_ac::WorkspaceAccessControlImpl;
